+++ order.service/src/main/java/com/pooranjoyb/order/service/Application.java
@@ -1,0 +1,13 @@
+package com.pooranjoyb.order.service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+}
