+++ order.service/src/main/java/com/pooranjoyb/order/service/exception/ErrorResponse.java
@@ -1,0 +1,6 @@
+package com.pooranjoyb.order.service.exception;
+
+import java.time.LocalDateTime;
+
+public record ErrorResponse(LocalDateTime timestamp, String message, String detail) {
+}
