+++ order.service/src/main/java/com/pooranjoyb/order/service/order/common/enums/OrderStatus.java
@@ -1,0 +1,8 @@
+package com.pooranjoyb.order.service.order.common.enums;
+
+public enum OrderStatus {
+    PENDING,
+    BOOKED,
+    CANCELLED,
+    FAILED
+}
