+++ order.service/src/main/java/com/pooranjoyb/order/service/order/controller/OrderController.java
@@ -1,0 +1,36 @@
+package com.pooranjoyb.order.service.order.controller;
+
+import com.pooranjoyb.order.service.order.dto.OrderRequestDto;
+import com.pooranjoyb.order.service.order.dto.OrderResponseDto;
+import com.pooranjoyb.order.service.order.service.OrderService;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/orders")
+public class OrderController {
+
+    @Autowired
+    private final OrderService orderService;
+
+    public OrderController(OrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @GetMapping()
+    public ResponseEntity<List<OrderResponseDto>> getOrders() {
+        List<OrderResponseDto> orders = orderService.getOrders();
+        return new ResponseEntity<>(orders, HttpStatus.OK);
+    }
+
+    @PostMapping
+    public ResponseEntity<OrderResponseDto> createOrder(@RequestBody @Valid OrderRequestDto requestDto) {
+        OrderResponseDto order = orderService.createOrder(requestDto);
+        return new ResponseEntity<>(order, HttpStatus.CREATED);
+    }
+}
