+++ order.service/src/main/java/com/pooranjoyb/order/service/order/service/OrderService.java
@@ -1,0 +1,12 @@
+package com.pooranjoyb.order.service.order.service;
+
+import com.pooranjoyb.order.service.order.dto.OrderRequestDto;
+import com.pooranjoyb.order.service.order.dto.OrderResponseDto;
+
+import java.util.List;
+
+public interface OrderService {
+
+    OrderResponseDto createOrder(OrderRequestDto orderRequestDto);
+    List<OrderResponseDto> getOrders();
+}
