+++ order.service/src/test/java/com/pooranjoyb/order/service/OrderStreamApplicationTests.java
@@ -1,0 +1,13 @@
+package com.pooranjoyb.order.service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrderStreamApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
