+++ product.service/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+spring.application.name=product.service
+
+# database properties
+spring.datasource.username=admin
+spring.datasource.password=admin
+spring.datasource.url=jdbc:postgresql://localhost:5435/order_stream
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
