+++ product.service/src/test/java/com/pooranjoyb/product/service/ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.pooranjoyb.product.service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
