--- src/main/java/com/pooranjoyb/order_stream/OrderStreamApplication.java
@@ -1,13 +1,0 @@
-package com.pooranjoyb.order_stream;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderStreamApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderStreamApplication.class, args);
-	}
-
-}
