--- src/main/java/com/pooranjoyb/order_stream/exception/ErrorResponse.java
@@ -1,6 +1,0 @@
-package com.pooranjoyb.order_stream.exception;
-
-import java.time.LocalDateTime;
-
-public record ErrorResponse(LocalDateTime timestamp, String message, String detail) {
-}
