--- src/main/java/com/pooranjoyb/order_stream/exception/GlobalExceptionHandler.java
@@ -1,50 +1,0 @@
-package com.pooranjoyb.order_stream.exception;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.time.LocalDateTime;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    private static final Logger LOG = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-    /**
-     * Handles Bean Validation errors
-     * Returns 400 Bad Request
-     */
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorResponse> handleValidationException(MethodArgumentNotValidException ex) {
-        String errorMessage = ex.getBindingResult().getFieldErrors().get(0).getDefaultMessage();
-        String fieldName = ex.getBindingResult().getFieldErrors().get(0).getField();
-        LOG.error(ex.getMessage());
-        ErrorResponse error = new ErrorResponse(
-                LocalDateTime.now(),
-                "Validation Failed",
-                fieldName + ": " + errorMessage
-        );
-
-        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
-    }
-
-    /**
-     * Catching for any other unhandled exceptions
-     * Returns 500 Internal Server Error
-     */
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleAllExceptions(Exception ex) {
-        LOG.error(ex.getMessage());
-        ErrorResponse error = new ErrorResponse(
-                LocalDateTime.now(),
-                "Internal Server Error",
-                "Something went wrong. Please try again later."
-        );
-
-        return new ResponseEntity<>(error, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
