--- src/main/java/com/pooranjoyb/order_stream/order/common/enums/OrderStatus.java
@@ -1,8 +1,0 @@
-package com.pooranjoyb.order_stream.order.common.enums;
-
-public enum OrderStatus {
-    PENDING,
-    BOOKED,
-    CANCELLED,
-    FAILED
-}
