--- src/main/java/com/pooranjoyb/order_stream/order/repository/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.pooranjoyb.order_stream.order.repository;
-
-import com.pooranjoyb.order_stream.order.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
