--- src/main/java/com/pooranjoyb/order_stream/order/service/OrderService.java
@@ -1,13 +1,0 @@
-package com.pooranjoyb.order_stream.order.service;
-
-import com.pooranjoyb.order_stream.order.dto.OrderRequestDto;
-import com.pooranjoyb.order_stream.order.dto.OrderResponseDto;
-import com.pooranjoyb.order_stream.order.entity.Order;
-
-import java.util.List;
-
-public interface OrderService {
-
-    Order createOrder(OrderRequestDto orderRequestDto);
-    List<OrderResponseDto> getOrders();
-}
