--- src/test/java/com/pooranjoyb/order_stream/OrderStreamApplicationTests.java
@@ -1,13 +1,0 @@
-package com.pooranjoyb.order_stream;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrderStreamApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
